Specs only delete test uploads if storage directory set

+++ spec/support/carrierwave.rb
@@ -6,9 +6,13 @@ Exportling.base_storage_directory = 'exportling_spec_directory'
 # Make sure we remove all uploaded files
 RSpec.configure do |config|
   config.before(:suite) do
-    FileUtils.rm_rf("#{Rails.root}/#{Exportling.base_storage_directory}")
+    if Exportling.base_storage_directory.present?
+      FileUtils.rm_rf("#{Rails.root}/#{Exportling.base_storage_directory}")
+    end
   end
   config.after(:suite) do
-    FileUtils.rm_rf("#{Rails.root}/#{Exportling.base_storage_directory}")
+    if Exportling.base_storage_directory.present?
+      FileUtils.rm_rf("#{Rails.root}/#{Exportling.base_storage_directory}")
+    end
   end
 end
